+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## (Unreleased)
+
+- [x] patch for manual prefixed at-rules. i.e `@-moz-document {...}`.
+- [x] patch for flexbox vendor prefix `flex-grow`, `flex-shrink` and `flex-basis`.
+- [ ] better handle special `:global` selector with `newlines` in cascade isolation mode.
+
 ## 3.4.0 October 20, 2017
 
 - patch incorrect vendor prefixing of justify-content